change screen folder path
