+++ lib/pages/home/screens/essay/widgets/essay_slide_widget.dart
@@ -3,7 +3,7 @@ import 'package:unknown_note_flutter/common/widgets/app_font.dart';
 import 'package:unknown_note_flutter/constants/gaps.dart';
 import 'package:unknown_note_flutter/constants/sizes.dart';
 import 'package:unknown_note_flutter/enums/enum_essay_category.dart';
-import 'package:unknown_note_flutter/screens/essay/widgets/essay_category_button.dart';
+import 'package:unknown_note_flutter/pages/home/screens/essay/widgets/essay_category_button.dart';
 
 class EssaySlideWidget extends StatelessWidget {
   const EssaySlideWidget({super.key});
